+++ apps/backend/src/Application/MemberFaces/Queries/SearchVectorFace/SearchMemberFaceQuery.cs
@@ -8,6 +8,6 @@ public record SearchMemberFaceQuery : IRequest<Result<List<FoundFaceDto>>>
     }
     public Guid FamilyId { get; init; } // Bắt buộc: lọc tìm kiếm theo một gia đình cụ thể
     public Guid? MemberId { get; init; } // Optional: filter search to a specific member
-    public List<double> Vector { get; init; } = new List<double>(); // Face embedding to search with
+    public List<float> Vector { get; init; } = new List<float>(); // Face embedding to search with
     public int Limit { get; init; } = 5; // Max number of results
 }
